backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
feedparser==6.0.10
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2
python-dateutil==2.8.2
